requirements.txt
mattermostdriver=6.1.3
jira=2.0.0
Pillow=6.0.0
ldap3=2.6
colorclass=2.2.0
docopt=0.6.2